新增骷髅人的移动逻辑和不完善的攻击逻辑

+++ Assets/Scripts/General/Attack.cs
@@ -7,7 +7,7 @@ public class Attack : MonoBehaviour
     [Header("»ù±¾ÊôĞÔ")]
     public int damage;      // ÉËº¦Öµ
     public float attackRange;     // ¹¥»÷·¶Î§
-    public int attackRate;      // ¹¥»÷ÆµÂÊ
+    public float attackRate;      // ¹¥»÷ÆµÂÊ
 
     private void OnTriggerStay2D(Collider2D other)
     {
